added typeDefs content
