+++ server/schemas/resolvers.js
@@ -1,36 +1,4 @@
-const typeDefs = `
+const typeDefs = require('./typeDefs');
+const resolvers = require('./resolvers');
 
-type Auth {
-    token: ID!
-    user: User
-  }
-
-type Book {
-    bookId
-    authors
-    description
-    title
-    image
-    link
-
-}
-
-type User {
-    _id: ID
-    username: String
-    email: String
-    bookCount
-    savedBooks
-}
-
-type Mutation {
-    login(email: String!, password: String!): Auth
-    addUser(firstName: String!, lastName: String!, email: String!, password: String!): Auth
-    saveBook
-    removeBook
-}
-
-type Query {
-    me: User
-  }
-  ;`
+module.exports = { typeDefs, resolvers };
